requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.7
django-allauth==0.44.0
django-filter==2.4.0
django-icons==2.2.1
django-import-export==2.5.0
Django-Verify-Email==0.0.5
et-xmlfile==1.0.1
idna==2.10
importlib-metadata==2.1.1
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.1.1
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
split==0.4
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
